+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 * (theimo1221) Allow Heating Rod even during winter
 * (theimo1221) Generalize Temperature Sensor
+* (theimo1221) Generalize Humidity Sensor
 
 ## 3.0.0-alpha.89 (2024-10-03)